systemd: bump to v213
This patch bumps systemd to v213.

This new version introduces systemd-timesyncd, a SNTP daemon. This
feature can be enabled if systemd-networkd is selected via the
configuration menu. It is a simple alternative to ntpd, useful for
machines without a RTC.

The patch for reverting the use of --relative option when calling `ln`
has been refreshed, as the configure script now checks if `ln` supports
the --relative option and fails if it does not.

Another patch for the proper deactivation of gtk-doc has been added.

All these steps now require an autoreconf.

Signed-off-by: Eric Le Bihan <eric.le.bihan.dev@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>