package/libcec: fix missing libraries from rpi-userland
When rpi-userland libraries are built as shared-objects, not all needed
libraries are passed in the LD_FLAGS, leading to failure at linkage.

To avoid this issue, set the LIBS variable content in accordance with
INCLUDES variable value.

This is needed because libcec does not use pkg-config to look for the
ldflags needed for -lbcm_host, even though bcm_host provides a .pc
file (even a correct one!)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Samuel Martin <s.martin49@gmail.com>
Tested-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>