freescale-imx: bump to version 3.10.17-1.0.0
Bumping the version requires some heavy changes in the different
i.MX related packages. Instead of having a single huge patch, those
changes are split in smaller patches, one per affected package.

This has the inconvenient side-effect of breaking the build right
now with the bump; but it is better since the patches are easier to
understand.

Signed-off-by: Gary Bisson <bisson.gary@gmail.com>
Acked-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
[yann.morin.1998@free.fr: expand commit log]
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Hadrien Boutteville <hadrien.boutteville@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>