freescale-imx: update imx-lib package
The include files from the FSL kernel are properly separated out in
uapi, so we must include that instead of the kernel-side headers.

imx-lib no longer includes VPU code and therefore is now free from
Freescale License. It is now LGPLv2.1+.

Also, it is now provided as a normal .tar.gz, so _SOURCE is no longer
needed. So, remove our custom _EXTRACT_CMDS.

Signed-off-by: Gary Bisson <bisson.gary@gmail.com>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
[yann.morin.1998@free.fr: _SOURCE is no longer needed, rework commit
 log]
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Hadrien Boutteville <hadrien.boutteville@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>