package/libfcgi: do not build examples
Examples do not compile for a static build, because libtool forcibly
wants to link against the shared version of libstdc++.so, instead of the
static version.

Since we do not really need the examples on the target, just disable
them, it is easier than fixing libtool.

Also, while we're at it, rename the patches to include a sequence
number.

Fixes:
    http://autobuild.buildroot.org/results/66e/66e19cfd5a6048d2aaa46365518ea875e0abcb4d/
    http://autobuild.buildroot.org/results/71d/71d9ba50b41efa05a3aff06c24412d5e8b81d8f1/

[Thomas: edit commit log to indicate that the patches are renamed to
include a sequence number.]

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>