flickcurl: adjust openssl dependency
flickcurl does not specifically need openssl: it needs libcurl to have
crypto support to access https:// URLs. So, instead of depending on
openssl in FLICKCURL_DEPENDENCIES, this commit simply leaves it to the
Config.in file to select either OpenSSL, gnutls or libnss to make sure
libcurl will use one of these libraries.

Reported-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>