libevdev: remove host python dependency
In commit 79c31e3297da7f0aec4d4675c154736e4cffe3ed ("package/libevdev:
add dependency on host-python"), Yann added a dependency of libevdev
on host-python, because the Python script in evdev was using argparse
which is only available since Python 2.6. At the time, this was
causing failures on machines that were using Python versions older
than 2.6.

However, since Yann's patch, the libevdev version has been bumped. And
one of the upstream commit,
http://cgit.freedesktop.org/libevdev/commit/libevdev/?id=cea8f0b8cc86332b0643acd62f24f9fef2ecc153,
was precisely made to remove the argparse dependency to avoid the
need to have a >= 2.6 Python version.

Thanks to this, we can avoid the host-python dependency and rely on
the system-installed Python version instead.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: "Yann E. MORIN" <yann.morin.1998@free.fr>