imagemagick: Disable OpenMP
By default (i.e. without adding libgomp.so* to
BR2_TOOLCHAIN_EXTRA_EXTERNAL_LIBS), if an external toolchain with libgomp
support is used, then libgomp is installed to staging/ , but not to target/ .
Consequently, with such a toolchain, imagemagick's configure detected libgomp as
supported and enabled its usage for the build stage, but then it failed to run
on the target because libgomp was missing. This is the bug #7322.

Disable OpenMP for imagemagick as a temporary workaround, until a better long
term solution is agreed on and implemented.

Signed-off-by: Benoît Thébaudeau <benoit.thebaudeau.dev@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>