directfb: disable multi application if no MMU
directfb-fusion uses the madvise() system call which is not available
on no-MMU targets.

It seems it might be possible to simply remove the call to madvise()
(there is only one call) if not available. However, it's probably not
worth the effort, and people working on no-MMU targets can submit a
patch doing that if they are interested.

[Thomas: tweak commit log.]

Signed-off-by: Jérôme Pouiller <jezz@sysmic.org>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>