ngrep: update include path of libpcap so bpf.h can be found
The libpcap headers have moved from /usr/include to /usr/include/pcap,
although /usr/include/pcap.h still exists for backwards-compatibility
(but is a stub that simply includes pcap/pcap.h).
The ngrep package tries to find bpf.h based on the include statements in
pcap.h. If we point ngrep to /usr/include/pcap.h, bpf.h will not be
found and the associated functionality will not work.

The fix is to pass the new include path to the ngrep configure step.

Fixes bug #7370

[1] https://bugs.busybox.net/show_bug.cgi?id=7370

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>